requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
openai>=1.14,<2
pyyaml==6.0.1
pytest==7.4.3
